Restructure logs components
