+++ apps/dashboard/src/entities/logs/models/LogObject.ts
@@ -1,0 +1,1 @@
+export type { LogObject } from 'consola'
