+++ apps/dashboard/src/entities/logs/ui/TagView.tsx
@@ -1,0 +1,3 @@
+export default function TagView({ tag }: { tag: string }) {
+  return <span className="text-gray-500">{tag}:</span>
+}
